move models into separate directory
